requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
amqp==1.4.8               # via kombu
anyjson==0.3.3            # via kombu
argparse==1.4.0           # via keystoneauth1, oslo.config, python-keystoneclient, python-novaclient, stevedore
astroid==1.4.3            # via pylint
babel==2.2.0              # via debtcollector, oslo.i18n, oslo.serialization, python-keystoneclient, python-novaclient
billiard==3.3.0.22        # via celery
celery==3.1.19
colorama==0.3.5           # via pylint
debtcollector==1.1.0      # via oslo.utils, python-keystoneclient
dj-database-url==0.3.0
django-admin-sortable==2.0.9
django-cumulus==1.0.19
django-mailgun==0.8.0
django-mptt==0.8.0
django-redis==4.3.0
django-simple-resizer==0.1.8
django-solo==1.1.1
django==1.9.1
feedparser==5.2.1         # via feincms
feincms==1.11.4
funcsigs==0.4             # via mock, oslo.utils
ip-associations-python-novaclient-ext==0.1  # via rackspace-novaclient
iso8601==0.1.11           # via keystoneauth1, oslo.utils, python-keystoneclient, python-novaclient
keyring==5.7.1            # via pyrax
keystoneauth1==2.1.0      # via python-keystoneclient
kombu==3.0.32             # via celery
lazy-object-proxy==1.2.1  # via astroid
markdown==2.6.5
mock==1.3.0               # via pyrax
monotonic==0.5            # via oslo.utils
msgpack-python==0.4.6     # via oslo.serialization
netaddr==0.7.18           # via oslo.config, oslo.utils, python-keystoneclient
netifaces==0.10.4         # via oslo.utils
newrelic==2.60.0.46
os-diskconfig-python-novaclient-ext==0.1.2  # via rackspace-novaclient
os-networksv2-python-novaclient-ext==0.25  # via rackspace-novaclient
os-virtual-interfacesv2-python-novaclient-ext==0.19  # via rackspace-novaclient
oslo.config==3.2.0        # via python-keystoneclient
oslo.i18n==3.1.0          # via oslo.utils, python-keystoneclient, python-novaclient
oslo.serialization==2.2.0  # via python-keystoneclient, python-novaclient
oslo.utils==3.3.0         # via oslo.serialization, python-keystoneclient, python-novaclient
pbr==1.8.1                # via debtcollector, mock, oslo.i18n, oslo.serialization, oslo.utils, python-keystoneclient, python-novaclient, stevedore
pep8==1.6.2
pillow==2.8.1
prettytable==0.7.2        # via python-keystoneclient, python-novaclient
psycopg2==2.6.1
pylint==1.5.2
pyrax==1.9.6
python-keystoneclient==2.0.0  # via python-novaclient
python-novaclient==2.27.0  # via ip-associations-python-novaclient-ext, os-diskconfig-python-novaclient-ext, os-networksv2-python-novaclient-ext, os-virtual-interfacesv2-python-novaclient-ext, pyrax, rackspace-auth-openstack, rackspace-novaclient, rax-default-network-flags-python-novaclient-ext, rax-scheduled-images-python-novaclient-ext
pytz==2015.7              # via babel, celery, feincms, oslo.serialization, oslo.utils
rackspace-auth-openstack==1.3  # via rackspace-novaclient
rackspace-novaclient==1.5  # via pyrax
rax-default-network-flags-python-novaclient-ext==0.3.2  # via rackspace-novaclient
rax-scheduled-images-python-novaclient-ext==0.3.1  # via rackspace-novaclient
redis==2.10.5             # via django-redis
requests==2.9.1           # via django-mailgun, keystoneauth1, pyrax, python-keystoneclient, python-novaclient
simplejson==3.8.1         # via python-novaclient
six==1.10.0               # via astroid, debtcollector, django-mailgun, keystoneauth1, mock, oslo.config, oslo.i18n, oslo.serialization, oslo.utils, pylint, pyrax, python-keystoneclient, python-novaclient, stevedore
stevedore==1.10.0         # via keystoneauth1, oslo.config, python-keystoneclient
uwsgi==2.0.12
Wand==0.4.1
wrapt==1.10.6             # via astroid, debtcollector
